restlight-ext/restlight-ext-filter/src/main/java/esa/restlight/ext/filter/accesslog/AccessLogOptions.java
/*
 * Copyright 2020 OPPO ESA Stack Project
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package esa.restlight.ext.filter.accesslog;

import java.io.Serializable;

public class AccessLogOptions implements Serializable {

    private static final long serialVersionUID = 5499602682942371890L;

    private String directory = "logs";
    private String fileName = "access.log";
    private String charset;
    private boolean rolling = true;
    private String datePattern = "yyyy-MM-dd";
    private int maxHistory = 10;
    private boolean fullUri;

    public String getDirectory() {
        return directory;
    }

    public void setDirectory(String directory) {
        this.directory = directory;
    }

    public String getFileName() {
        return fileName;
    }

    public void setFileName(String fileName) {
        this.fileName = fileName;
    }

    public boolean isRolling() {
        return rolling;
    }

    public void setRolling(boolean rolling) {
        this.rolling = rolling;
    }

    public String getCharset() {
        return charset;
    }

    public void setCharset(String charset) {
        this.charset = charset;
    }

    public String getDatePattern() {
        return datePattern;
    }

    public void setDatePattern(String datePattern) {
        this.datePattern = datePattern;
    }

    public int getMaxHistory() {
        return maxHistory;
    }

    public void setMaxHistory(int maxHistory) {
        this.maxHistory = maxHistory;
    }

    public boolean isFullUri() {
        return fullUri;
    }

    public void setFullUri(boolean fullUri) {
        this.fullUri = fullUri;
    }
}
